+++ src/Basic/Vector.cpp
@@ -524,14 +524,14 @@ VectorDouble ut_vector_sequence(double valFrom, double valTo, double valStep)
 int ut_vector_size(const VectorInt &vec)
 {
   if (vec.empty()) return 0;
-  int size = sizeof(std::vector<int>) + (sizeof(int) * vec.size());
+  int size = sizeof(std::vector<int>) + (sizeof(int) * (int) vec.size());
   return size;
 }
 
 int ut_vector_size(const VectorDouble &vec)
 {
   if (vec.empty()) return 0;
-  int size = sizeof(std::vector<double>) + (sizeof(double) * vec.size());
+  int size = sizeof(std::vector<double>) + (sizeof(double) * (int) vec.size());
   return size;
 }
 
@@ -540,7 +540,7 @@ int ut_vector_size(const VectorVectorInt &vec)
   int size = 0;
   if (vec.empty()) return size;
   for (auto i = 0; i != (int) vec.size(); i++)
-    size += sizeof(std::vector<int>) + (sizeof(int) * vec[i].size());
+    size += sizeof(std::vector<int>) + (sizeof(int) * (int) vec[i].size());
   return size;
 
 }
@@ -550,7 +550,7 @@ int ut_vector_size(const VectorVectorDouble &vec)
   int size = 0;
   if (vec.empty()) return size;
   for (auto i = 0; i != (int) vec.size(); i++)
-    size += sizeof(std::vector<double>) + (sizeof(double) * vec[i].size());
+    size += sizeof(std::vector<double>) + (sizeof(double) * (int) vec[i].size());
   return size;
 }
